+++ src/main/java/com/agroconnect/frogger/entity/Payment.java
@@ -11,7 +11,7 @@ public class Payment {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private BigInteger id;
 
     @OneToOne
     @JoinColumn(name = "orderId", nullable = false)
@@ -35,14 +35,14 @@ public class Payment {
     private Timestamp createdAt = new Timestamp(System.currentTimeMillis());
 
     // Getters and Setters
-    public Long getId() { return id; }
-    public void setId(Long id) { this.id = id; }
+    public BigInteger getId() { return id; }
+    public void setId(BigInteger id) { this.id = id; }
 
-    public Long getOrderId() { return order.getId(); }
-    public void setOrderId(Long id) { this.order.setId(id); }
+    public BigInteger getOrderId() { return order.getId(); }
+    public void setOrderId(BigInteger id) { this.order.setId(id); }
 
     public BigInteger getCustomerId() { return customer.getId(); }
-    public void setCustomerId(Long id) { this.customer.setId(BigInteger.valueOf(id)); }
+    public void setCustomerId(BigInteger id) { this.customer.setId(id); }
 
     public Double getTotalAmount() { return amount; }
     public void setTotalAmount(Double amount) { this.amount = amount; }
